Rename DSL services files and endpoints
