+++ GUI/src/pages/ConnectionRequests/index.tsx
