+++ GUI/src/services/requests.ts
@@ -1,8 +1,8 @@
-import api from './api-dev';
+import api from './api';
 import { Trigger } from 'types/trigger';
 
 export async function updateConnectionRequest(request: Trigger, status: "approved" | "declined") {
-  const { data } = await api.post<{ response: Trigger[]}>(`/services/respond-to-connection-request`, {
+  const { data } = await api.post<{ response: Trigger[]}>(`services/respond-to-connection-request`, {
     serviceId: request.service,
     serviceName: request.serviceName,
     intent: request.intent,
@@ -20,6 +20,6 @@ export interface RequestServiceIntentConnectionData {
 }
 
 export async function requestServiceIntentConnection(data: RequestServiceIntentConnectionData) {
-  const { data: result } = await api.post<Trigger>(`/services/request-service-intent-connection`, data);
+  const { data: result } = await api.post<Trigger>(`services/request-service-intent-connection`, data);
   return result;
 }
